+++ api/database/seeders/PoolRandomSeeder.php
@@ -3,29 +3,10 @@
 namespace Database\Seeders;
 
 use App\Models\Pool;
-use Faker\Generator;
-use Illuminate\Container\Container;
 use Illuminate\Database\Seeder;
 
-class PoolSeederRandom extends Seeder
+class PoolRandomSeeder extends Seeder
 {
-    /**
-     * The current Faker instance.
-     *
-     * @var \Faker\Generator
-     */
-    protected $faker;
-
-    /**
-     * Create a new seeder instance.
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-        $this->faker = Container::getInstance()->make(Generator::class);
-    }
-
     /**
      * Run the database seeds.
      *
