+++ api/database/seeders/ProdSeeder.php
@@ -8,15 +8,15 @@ class ProdSeeder extends Seeder
 {
     /**
      * Run the database seeds.
-     *
-     * @return void
      */
-    public function run()
+    public function run(): void
     {
         $this->call([
+            // standard platform data
             RolePermissionSeeder::class,
             ClassificationSeeder::class,
             DepartmentSeeder::class,
+            GenericJobTitleSeeder::class,
             SkillFamilySeeder::class,
             SkillSeeder::class,
             TeamSeeder::class,
