+++ api/database/seeders/TeamSeeder.php
@@ -6,7 +6,7 @@ use App\Models\Department;
 use App\Models\Team;
 use Illuminate\Database\Seeder;
 
-class TeamTestSeeder extends Seeder
+class TeamSeeder extends Seeder
 {
     /**
      * Run the database seeds.
@@ -22,7 +22,7 @@ class TeamTestSeeder extends Seeder
                     'en' => 'Digital Community Management',
                     'fr' => 'Gestion de la collectivité numérique',
                 ],
-                'contact_email' => 'dcm@test.test',
+                'contact_email' => 'gctalent-talentgc@support-soutien.gc.ca',
                 'department_ids' => [Department::select('id')->where('name->en', 'Treasury Board Secretariat')->sole()->id],
             ],
             [
@@ -31,7 +31,7 @@ class TeamTestSeeder extends Seeder
                     'en' => 'Office of Indigenous Initiatives',
                     'fr' => 'Bureau initiatives autochtones',
                 ],
-                'contact_email' => 'oit@test.test',
+                'contact_email' => 'edsc.pda-iap.esdc@hrsdc-rhdcc.gc.ca',
                 'department_ids' => [Department::select('id')->where('name->en', 'Employment and Social Development (Department of)')->sole()->id],
             ],
         ];
